% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL_Combine.R
\name{FL_combine}
\alias{FL_combine}
\title{FL_Combine}
\usage{
FL_combine(summaries)
}
\arguments{
\item{summaries}{the outputs from the FL_local_summary which contains all local level summary statistics that will be used in this analysis}
}
\value{
the combined list of summary statistics is returned
}
\description{
This function will take all the local server summaries and combine them in preparation for
the federated learning analysis.  It requires all the local server summaries as an input, these
will be exactly what the R Shiny app produces.
}
\examples{
data(mtcars)
mk_lm(Y = mtcars$mpg, X = cbind(mtcars$wt, mtcars$cyl), X_names = c("weight", "cylinders"))
mk_lm(Y = mtcars$mpg, X = matrix(mtcars$wt, ncol = 1), X_names = c("weight"))

}
