% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL_Local_Summary.R
\name{FL_local_summary}
\alias{FL_local_summary}
\title{FL_Local_Summary}
\usage{
FL_local_summary(formula, data)
}
\arguments{
\item{formula}{linear regression formula being used for the study}

\item{data}{csv file containing all the specified data fro the study.}
}
\value{
a list containing the summary statistics (XtX matrix, XtY matrix, YtY matrix, & n) is returned
}
\description{
This function will produce the local server federated learning summary statistics.  It requires an input of the
linear regression formula and a csv data set.  It will then run a linear regression and output the necessary
summary statistics (XtX matrix, XtY matrix, YtY matrix, & n)
}
\examples{
data(mtcars)
mk_lm(Y = mtcars$mpg, X = cbind(mtcars$wt, mtcars$cyl), X_names = c("weight", "cylinders"))
mk_lm(Y = mtcars$mpg, X = matrix(mtcars$wt, ncol = 1), X_names = c("weight"))

}
