% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthcare_dataset}
\alias{healthcare_dataset}
\title{My Example Dataset}
\description{
This is an example dataset included in the FinalProject625 package.  It is a simulated dataset from Kaggle
}
